resume_parsing/requirements.txt
streamlit
python-docx
requests
PyPDF2
docx2txt 